poly_int: TYPE_VECTOR_SUBPARTS
This patch changes TYPE_VECTOR_SUBPARTS to a poly_uint64.  The value is
encoded in the 10-bit precision field and was previously always stored
as a simple log2 value.  The question was how the 10 bits should be
used for variable-length vectors.  In practice the number of vector
elements should always have the form N + N * X (where X is the runtime
value), and as for constant-length vectors, N must be a power of 2.
The patch therefore uses the low bit to select between constant-length
and variable-length and uses the upper 9 bits to encode log2(N).
Targets without variable-length vectors continue to use the old scheme.
A new valid_vector_subparts_p function tests whether a given number of
elements can be encoded.

The patch disables various things that won't work yet for
variable-length vectors.  Later patches relax some of the conditions.

git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ARM@242271 138bc75d-0d04-0410-961f-82ee72b054a4

+++ sve-branch/gcc/config/rs6000/rs6000.c
@@ -5379,7 +5379,7 @@ rs6000_builtin_vectorization_cost (enum vect_cost_for_stmt type_of_cost,
 	    && TYPE_PRECISION (elem_type) == 32)
 	  return 5;
 	else
-	  return max (2, TYPE_VECTOR_SUBPARTS (vectype) - 1);
+	  return max (2, (int) TYPE_VECTOR_SUBPARTS (vectype) - 1);
 
       default:
         gcc_unreachable ();
